+++ packages/core/src/components/Pagination/PaginationPopup/index.ts
@@ -1,0 +1,3 @@
+import { PaginationPopup } from './PaginationPopup';
+
+export default PaginationPopup;