+++ packages/core/src/components/Pagination/PaginationPopup/types.ts
@@ -1,0 +1,5 @@
+export interface Props {
+    prevPageNumber: number;
+    nextPageNumber: number;
+    onClickHandler: (page: number) => void;
+}